Adapt imports and exports
