+++ experimental/LieAlgebras/src/LieAlgebras.jl
@@ -1,10 +1,14 @@
-isdefined(Oscar, :word) || function word end
-
 module LieAlgebras
 
 using ..Oscar
 
-import Oscar: GAPWrap, IntegerUnion, MapHeader
+using Oscar:
+  _root_system_type_string,
+  _vec,
+  GAPWrap,
+  IntegerUnion,
+  MapHeader,
+  set_root_system_type!
 
 import Random
 
@@ -19,20 +23,19 @@ import ..Oscar:
   _is_exterior_power,
   _is_tensor_product,
   _iso_oscar_gap,
-  _vec,
   action,
-  add!,
-  addmul!,
   basis_matrix,
   basis,
   canonical_injection,
   canonical_injections,
   canonical_projection,
   canonical_projections,
+  cartan_matrix,
   center,
   centralizer,
   character,
   characteristic,
+  check_parent,
   coeff,
   coefficient_ring,
   coefficients,
@@ -41,7 +44,6 @@ import ..Oscar:
   derived_series,
   dim,
   direct_sum,
-  dot,
   dual,
   elem_type,
   expressify,
@@ -49,7 +51,6 @@ import ..Oscar:
   fp_group,
   gen,
   gens,
-  height,
   hom,
   hom_direct_sum,
   hom_tensor,
@@ -59,8 +60,6 @@ import ..Oscar:
   induced_map_on_exterior_power,
   inv,
   is_abelian,
-  is_finite,
-  is_gen,
   is_isomorphism,
   is_nilpotent,
   is_perfect,
@@ -71,25 +70,17 @@ import ..Oscar:
   kernel,
   lower_central_series,
   matrix,
-  mul!,
-  neg!,
   normalizer,
   number_of_generators,
   ngens,
-  order,
   parent_type,
   permutation_group,
   rank,
-  root,
-  roots,
+  root_system,
   sub,
-  sub!,
   symbols,
   symmetric_power,
   tensor_product,
-  weyl_vector,
-  word,
-  zero!,
   zero_map,
   ⊕,
   ⊗
@@ -109,15 +100,6 @@ export _is_standard_module
 export _is_symmetric_power
 export _is_tensor_power
 export _is_tensor_product
-
-# Aliases
-function number_of_positive_roots end
-function number_of_roots end
-function number_of_simple_roots end
-
-@alias n_positive_roots number_of_positive_roots
-@alias n_roots number_of_roots
-@alias n_simple_roots number_of_simple_roots
 
 include("Types.jl")
 include("Combinatorics.jl")
