+++ experimental/LieAlgebras/src/exports.jl
@@ -4,7 +4,6 @@
 
 export AbstractLieAlgebra, AbstractLieAlgebraElem
 export DirectSumLieAlgebra, DirectSumLieAlgebraElem
-export DualRootSpaceElem
 export LieAlgebra, LieAlgebraElem
 export LieAlgebraHom
 export LieAlgebraIdeal
@@ -12,10 +11,6 @@ export LieAlgebraModule, LieAlgebraModuleElem
 export LieAlgebraModuleHom
 export LieSubalgebra
 export LinearLieAlgebra, LinearLieAlgebraElem
-export RootSpaceElem
-export RootSystem
-export WeightLattice, WeightLatticeElem
-export WeylGroup, WeylGroupElem
 
 export abelian_lie_algebra
 export abstract_module
@@ -23,18 +18,9 @@ export adjoint_matrix
 export any_non_ad_nilpotent_element
 export base_lie_algebra
 export bracket
-export cartan_bilinear_form
-export cartan_matrix
 export cartan_subalgebra
-export cartan_symmetrizer
-export cartan_type
-export cartan_type_with_ordering
 export chevalley_basis
 export coerce_to_lie_algebra_elem
-export conjugate_dominant_weight
-export conjugate_dominant_weight_with_elem
-export coroot
-export coroots
 export coxeter_matrix
 export demazure_character
 export demazure_operator
@@ -44,61 +30,18 @@ export dominant_character
 export dominant_weights
 export engel_subalgebra
 export exterior_power
-export fundamental_weight
-export fundamental_weights
 export general_linear_lie_algebra
+export has_root_system
 export induced_map_on_symmetric_power
 export induced_map_on_tensor_power
 export is_ad_nilpotent
-export is_cartan_matrix
-export is_cartan_type
-export is_coroot
-export is_coroot_with_index
-export is_dominant
-export is_fundamental_weight
-export is_fundamental_weight_with_index
-export is_negative_coroot
-export is_negative_coroot_with_index
-export is_negative_root
-export is_negative_root_with_index
-export is_positive_coroot
-export is_positive_coroot_with_index
-export is_positive_root
-export is_positive_root_with_index
-export is_root
-export is_root_with_index
 export is_self_normalizing
-export is_simple_coroot
-export is_simple_coroot_with_index
-export is_simple_root
-export is_simple_root_with_index
 export killing_matrix
 export lie_algebra
-export longest_element
 export lower_central_series
 export matrix_repr_basis
-export negative_coroot
-export negative_coroots
-export negative_root
-export negative_roots
-export number_of_positive_roots
-export number_of_roots
-export number_of_simple_roots
-export positive_coroot
-export positive_coroots
-export positive_root
-export positive_roots
-export reduced_expressions
-export reflect, reflect!
-export root_system, has_root_system
-export root_system_type, has_root_system_type
-export root_system_type_with_ordering
 export show_dynkin_diagram
-export simple_coroot
-export simple_coroots
 export simple_module
-export simple_root
-export simple_roots
 export special_linear_lie_algebra
 export special_orthogonal_lie_algebra
 export standard_module
@@ -108,7 +51,3 @@ export tensor_power
 export tensor_product_decomposition
 export trivial_module
 export universal_enveloping_algebra
-export weight_lattice
-export weyl_group
-export weyl_orbit
-export word
