+++ src/exports.jl
@@ -65,6 +65,7 @@ export CyclicQuotientSingularity
 export DirectProductGroup
 export DirectSumSheaf
 export Directed
+export DualRootSpaceElem
 export Edge
 export EffectiveCartierDivisor
 export EllipticSurface
@@ -154,6 +155,8 @@ export QQ
 export R10_matroid
 export RationalEquivalenceClass
 export RayVector
+export RootSpaceElem
+export RootSystem
 export SIM_body_polytope
 export SL
 export SLPoly
@@ -193,7 +196,9 @@ export Undirected
 export VarietyFunctionField
 export VarietyFunctionFieldElem
 export WeakComposition
+export WeightLattice, WeightLatticeElem
 export WeilDivisor
+export WeylGroup, WeylGroupElem
 export WreathProductGroup
 export YoungTableau
 export ZZ
@@ -321,6 +326,11 @@ export canonical_divisor
 export canonical_divisor_class
 export canonical_isomorphism
 export canonical_matrix
+export cartan_bilinear_form
+export cartan_matrix
+export cartan_symmetrizer
+export cartan_type
+export cartan_type_with_ordering
 export cartesian_power
 export cartier_divisor
 export catalan_solid
@@ -415,6 +425,8 @@ export conjugacy_class
 export conjugacy_classes
 export conjugacy_classes_subgroups
 export conjugate
+export conjugate_dominant_weight, conjugate_dominant_weight!
+export conjugate_dominant_weight_with_elem, conjugate_dominant_weight_with_elem!
 export conjugate_group
 export conjugate_transpose
 export connected_components
@@ -436,6 +448,8 @@ export coordinates
 export cophenetic_matrix
 export corank
 export core
+export coroot
+export coroots
 export corresponding_bilinear_form
 export corresponding_quadratic_form
 export cotangent_sheaf
@@ -607,6 +621,8 @@ export fundamental_circuit
 export fundamental_cocircuit
 export fundamental_group
 export fundamental_invariants
+export fundamental_weight
+export fundamental_weights
 export g_vector
 export galois_group
 export galois_ideal
@@ -794,6 +810,8 @@ export is_binomial
 export is_bounded
 export is_canonically_isomorphic
 export is_canonically_isomorphic_with_map
+export is_cartan_matrix
+export is_cartan_type
 export is_cartier
 export is_cellular
 export is_character_table_name
@@ -809,10 +827,13 @@ export is_conjugate_subgroup
 export is_conjugate_subgroup_with_data
 export is_conjugate_with_data
 export is_connected
+export is_coroot
+export is_coroot_with_index
 export is_cyclic, has_is_cyclic, set_is_cyclic
 export is_degenerate
 export is_dense
 export is_dihedral_group, has_is_dihedral_group, set_is_dihedral_group
+export is_dominant
 export is_du_val_singularity
 export is_duplicate_table
 export is_effective
@@ -836,6 +857,8 @@ export is_full_direct_product
 export is_full_semidirect_product
 export is_full_wreath_product
 export is_fulldimensional
+export is_fundamental_weight
+export is_fundamental_weight_with_index
 export is_geometrically_integral
 export is_geometrically_reduced
 export is_global
@@ -876,6 +899,10 @@ export is_modular
 export is_natural_alternating_group, has_is_natural_alternating_group, set_is_natural_alternating_group
 export is_natural_symmetric_group, has_is_natural_symmetric_group, set_is_natural_symmetric_group
 export is_nef
+export is_negative_coroot
+export is_negative_coroot_with_index
+export is_negative_root
+export is_negative_root_with_index
 export is_nilpotent, has_is_nilpotent, set_is_nilpotent
 export is_non_zero_divisor
 export is_normal
@@ -889,6 +916,10 @@ export is_pgroup, has_is_pgroup, set_is_pgroup
 export is_pgroup_with_prime
 export is_platonic_solid
 export is_pointed
+export is_positive_coroot
+export is_positive_coroot_with_index
+export is_positive_root
+export is_positive_root_with_index
 export is_positively_graded
 export is_primary
 export is_prime
@@ -909,11 +940,17 @@ export is_reduced
 export is_regular
 export is_regular_sequence
 export is_right
+export is_root
+export is_root_with_index
 export is_semiregular
 export is_semisimple
 export is_semistandard
 export is_shifted
 export is_simple, has_is_simple, set_is_simple
+export is_simple_coroot
+export is_simple_coroot_with_index
+export is_simple_root
+export is_simple_root_with_index
 export is_simplicial
 export is_singular
 export is_smooth
@@ -1015,6 +1052,7 @@ export load_lp
 export load_mps
 export localization
 export localized_ring
+export longest_element
 export loops
 export low_index_subgroup_classes
 export low_index_subgroups
@@ -1126,6 +1164,10 @@ export name
 export names_of_fusion_sources
 export natural_character
 export nef_cone
+export negative_coroot
+export negative_coroots
+export negative_root
+export negative_roots
 export negbias
 export negdeglex
 export negdegrevlex
@@ -1173,7 +1215,10 @@ export number_of_moved_points, has_number_of_moved_points, set_number_of_moved_p
 export number_of_partitions
 export number_of_patches
 export number_of_perfect_groups, has_number_of_perfect_groups
+export number_of_positive_roots
 export number_of_primitive_groups, has_number_of_primitive_groups
+export number_of_roots
+export number_of_simple_roots
 export number_of_small_groups, has_number_of_small_groups
 export number_of_standard_tableaux
 export number_of_transitive_groups, has_number_of_transitive_groups
@@ -1258,7 +1303,11 @@ export polyhedral_fan_from_rays_action
 export polyhedron
 export polynomial
 export polynomial_ring
+export positive_coroot
+export positive_coroots
 export positive_hull
+export positive_root
+export positive_roots
 export possible_class_fusions
 export power_sum
 export powers_of_element
@@ -1345,9 +1394,11 @@ export reduce_fraction
 export reduce_with_quotients
 export reduce_with_quotients_and_unit
 export reduced_characteristic_polynomial
+export reduced_expressions
 export reduced_groebner_basis
 export reduced_scheme
 export reducible_fibers
+export reflect, reflect!
 export register_morphism!
 export regular_120_cell
 export regular_24_cell
@@ -1396,6 +1447,9 @@ export ring_elem_type
 export ring_type
 export rising_factorial
 export root
+export root_system
+export root_system_type, has_root_system_type
+export root_system_type_with_ordering
 export roots
 export row
 export rss_associahedron
@@ -1447,6 +1501,10 @@ export show_morphism_as_map
 export show_subquo
 export signed_incidence_matrix
 export signed_permutahedron
+export simple_coroot
+export simple_coroots
+export simple_root
+export simple_roots
 export simplex
 export simplicial_complex
 export simplicial_product
@@ -1626,13 +1684,17 @@ export weierstrass_contraction
 export weierstrass_model
 export weight
 export weight_cone
+export weight_lattice
 export weight_ordering
 export weighted_projective_space
 export weil_divisor
 export weyl_algebra
+export weyl_group
+export weyl_orbit
 export weyl_vector
 export with_ordering
 export witt_index
+export word
 export wreath_product
 export write_as_full
 export young_tableau
