+++ src/imports.jl
@@ -56,6 +56,7 @@ import AbstractAlgebra:
   @attributes,
   @show_name,
   @show_special,
+  @show_special_elem,
   allow_unicode,
   base_ring,
   canonical_unit,
